fix: consolidate authentication step definitions and fix BDD test imports
- Consolidated multiple authentication-related step definitions into single authentication_steps.js file
- Added backward compatibility for username-based authentication steps
- Fixed missing .js extensions in ES6 import statements across all step definitions
- Updated powers_steps.js to use Then instead of Given for result verification
- Fixed import paths in support files to include .js extensions

This resolves BDD test execution issues and enables the powers feature tests to run correctly.
