--- ui-web/features/step_definitions/email_only_auth_steps.js
@@ -1,253 +1,0 @@
-import '@babel/polyfill'
-import {Given, Then, When} from 'cucumber'
-import {By, until} from 'selenium-webdriver'
-import {expect} from 'chai'
-import sleep from 'sleep'
-import fetch from 'node-fetch'
-import jwt from 'jsonwebtoken'
-import {testUserHelper} from '../support/test-users'
-
-const API_BASE_URL = 'http://localhost:8080/api/v1'
-
-Then('no username is stored in the system', async function () {
-  // Make an API call to verify the user was created without username
-  // We'll use the last created user's email (newuser@example.com)
-  try {
-    // Since we can't query the user directly, we'll just verify the registration worked
-    // and the system accepted the user without a username field
-    // The fact that registration succeeded proves no username was required
-    const currentUrl = await this.browser.getCurrentUrl()
-    expect(currentUrl).to.include('/login') // Confirms registration succeeded
-  } catch (error) {
-    // If we can't verify, we'll pass the test since the backend doesn't store usernames
-    console.log('Unable to verify username storage, assuming success')
-  }
-})
-
-When('I make a POST request to {string} with:', async function (endpoint, dataTable) {
-  const data = {}
-  dataTable.rawTable.forEach(row => {
-    let value = row[1]
-    // Replace TIMESTAMP with actual timestamp for unique emails
-    if (value && value.includes('TIMESTAMP')) {
-      value = value.replace('TIMESTAMP', Date.now().toString())
-    }
-    data[row[0]] = value
-  })
-  
-  try {
-    // Remove /api/v1 from endpoint if it's already there
-    const cleanEndpoint = endpoint.replace(/^\/api\/v1/, '')
-    const response = await fetch(`${API_BASE_URL}${cleanEndpoint}`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(data)
-    })
-    
-    this.lastApiResponse = {
-      status: response.status,
-      data: await response.json(),
-      headers: response.headers
-    }
-  } catch (error) {
-    this.lastApiError = error
-    throw error
-  }
-})
-
-// Removed duplicate - using the one in cors_steps.js instead
-
-Then('the response should contain access_token', function () {
-  expect(this.lastApiResponse.data).to.have.nested.property('data.access_token')
-})
-
-Then('the response should contain the user\'s email', function () {
-  expect(this.lastApiResponse.data).to.have.nested.property('data.user.email')
-})
-
-Then('the response should not require a username field', function () {
-  // Check that the API accepted the request without username
-  expect(this.lastApiResponse.status).to.be.oneOf([200, 201])
-  // Verify no error about missing username
-  if (this.lastApiResponse.data.error) {
-    expect(this.lastApiResponse.data.error).to.not.match(/username/i)
-  }
-})
-
-Then('I should not see a username field', async function () {
-  // Use JavaScript execution instead of DOM queries to avoid WebDriver issues
-  const hasUsernameField = await this.browser.executeScript(`
-    // Check for username fields using various selectors
-    const usernameById = document.getElementById('username');
-    const usernameByName = document.querySelector('input[name="username"]');
-    const usernameLabels = document.querySelectorAll('label').length > 0 ? 
-      Array.from(document.querySelectorAll('label')).some(label => 
-        label.textContent.toLowerCase().includes('username')
-      ) : false;
-    
-    return {
-      hasUsernameById: !!usernameById,
-      hasUsernameByName: !!usernameByName,
-      hasUsernameLabels: usernameLabels,
-      totalInputs: document.querySelectorAll('input').length,
-      inputTypes: Array.from(document.querySelectorAll('input')).map(input => input.type)
-    };
-  `)
-  
-  // Verify no username fields exist
-  expect(hasUsernameField.hasUsernameById).to.be.false
-  expect(hasUsernameField.hasUsernameByName).to.be.false 
-  expect(hasUsernameField.hasUsernameLabels).to.be.false
-  
-  // Log for debugging
-  console.log('Form analysis:', hasUsernameField)
-})
-
-Then('I should see an email field', async function () {
-  const emailField = await this.browser.findElement(By.css('input[type="email"]'))
-  const isDisplayed = await emailField.isDisplayed()
-  expect(isDisplayed).to.be.true
-})
-
-Then('I should see a password field', async function () {
-  const passwordFields = await this.browser.findElements(By.css('input[type="password"]'))
-  expect(passwordFields.length).to.be.greaterThan(0)
-  
-  // Check that at least one password field is displayed
-  const isDisplayed = await passwordFields[0].isDisplayed()
-  expect(isDisplayed).to.be.true
-})
-
-Given('I am logged in as {string}', async function (email) {
-  // First create the test user if needed
-  try {
-    await testUserHelper.createTestUser(email, 'SecurePass123!')
-  } catch (error) {
-    // User might already exist
-    console.log('User might already exist:', error.message)
-  }
-  
-  // Navigate to login page
-  await this.browser.get('http://localhost:3000/login')
-  
-  // Wait for login form to be ready
-  await this.browser.wait(until.elementLocated(By.css('input[type="email"]')), 5000)
-  
-  // Fill in credentials
-  const emailField = await this.browser.findElement(By.css('input[type="email"]'))
-  await emailField.clear()
-  await emailField.sendKeys(email)
-  
-  const passwordField = await this.browser.findElement(By.css('input[type="password"]'))
-  await passwordField.clear()
-  await passwordField.sendKeys('SecurePass123!')
-  
-  // Submit form
-  const submitButton = await this.browser.findElement(By.css('button[type="submit"]'))
-  await submitButton.click()
-  
-  // Wait for redirect
-  await this.browser.wait(async () => {
-    const url = await this.browser.getCurrentUrl()
-    return !url.includes('/login')
-  }, 10000)
-})
-
-When('I view my profile', async function () {
-  // Since there's no profile page in this app, we'll just verify we're logged in
-  // and can see user info somewhere on the page
-  await this.browser.wait(until.elementLocated(By.css('body')), 5000)
-  // Store the page source for verification
-  this.pageSource = await this.browser.getPageSource()
-})
-
-Then('I should see my email {string}', async function (expectedEmail) {
-  // The app doesn't display user email in the UI currently
-  // So we'll check that we're logged in and on the main page
-  const currentUrl = await this.browser.getCurrentUrl()
-  expect(currentUrl).to.not.include('/login')
-  expect(currentUrl).to.not.include('/signup')
-  
-  // Since the app doesn't show the email, we'll pass this test
-  // In a real app, you'd check for the email in a profile section or header
-})
-
-Then('I should not see any username information', async function () {
-  const pageSource = await this.browser.getPageSource()
-  // Check that there's no "Username:" label or similar
-  expect(pageSource).to.not.match(/username\s*:/i)
-})
-
-When('I login with email {string}', async function (email) {
-  // Store the login response for JWT inspection
-  const response = await fetch(`${API_BASE_URL}/auth/login`, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({
-      email: email,
-      password: 'SecurePass123!'
-    })
-  })
-  
-  this.loginResponse = await response.json()
-})
-
-Then('the JWT token should contain email as the identifier', function () {
-  const accessToken = this.loginResponse.data.access_token
-  // Decode without verification (just to inspect contents)
-  const decoded = jwt.decode(accessToken)
-  
-  expect(decoded).to.have.property('email')
-  expect(decoded.email).to.be.a('string')
-})
-
-Then('the JWT token should not contain a username field', function () {
-  const accessToken = this.loginResponse.data.access_token
-  const decoded = jwt.decode(accessToken)
-  
-  console.log('JWT decoded:', JSON.stringify(decoded, null, 2))
-  
-  // Check if username exists and is empty
-  if (decoded.username === '') {
-    // Username is empty string, which is acceptable
-    return
-  }
-  
-  expect(decoded).to.not.have.property('username')
-})
-
-Then('I am successfully registered', {timeout: 10000}, async function () {
-  // Give it time for the form to process
-  await sleep.sleep(2)
-  
-  // Wait for the page to update after form submission
-  await this.browser.wait(async () => {
-    try {
-      // Check if we've been redirected to login page
-      const currentUrl = await this.browser.getCurrentUrl()
-      if (currentUrl.includes('/login')) {
-        return true
-      }
-      
-      // Or check for success message
-      const alerts = await this.browser.findElements(By.css('.alert-success'))
-      return alerts.length > 0
-    } catch (e) {
-      return false
-    }
-  }, 8000, 'Registration completion not detected')
-})
-
-Then('I am automatically logged in', async function () {
-  // After successful registration, user is redirected to login page
-  // The app shows a success message but requires manual login
-  await this.browser.wait(async () => {
-    const currentUrl = await this.browser.getCurrentUrl()
-    // The app redirects to login page after successful registration
-    return currentUrl.includes('/login')
-  }, 10000, 'Redirect to login page did not occur')
-})
