+++ ui-web/ui-web/features/support/BasicRules.js
@@ -1,0 +1,6 @@
+export default class BasicRules {
+	maximumNumberAttributePoints = 5
+	maximumNumberMajorHindrances = 1
+	maximumNumberMinorHindrances = 2
+	maximumNumberSkillPoints     = 15
+}
