+++ ui-web/ui-web/features/support/SettingRule.js
@@ -1,0 +1,4 @@
+export default class SettingRule {
+	name        = ''
+	description = ''
+}
