+++ ui-web/ui-web/features/support/chai.js
@@ -1,0 +1,8 @@
+// support/chai.js
+
+"use strict"
+
+let chai = require('chai')
+
+global.expect = chai.expect
+global.assert = chai.assert
